\name{couch_store_default}
\alias{couch_store_default}
\title{Store a document on the default connection.}
\usage{
couch_store_default(obj, myOpts = NULL)
}
\arguments{
  \item{obj}{A list formatted by calling couch_new_object.}

  \item{myOpts}{Additional options (not implemented in this
  version)}
}
\description{
Stores a record on the default connection.
}

